// src/types.ts
export type LocalWave = {
  id: number
  title: string
  type: string
  direction: string
  color: string
  opacity: number
  createdAt: string
  updatedAt: string
}

export type LocalBoxShadow = {
  id: number
  title: string
  color: string
  offsetX: number
  offsetY: number
  blurRadius: number
  spreadRadius: number
  shadowColor: string
  createdAt: string
  updatedAt: string
}